+++ phpunit/class-wp-theme-json-test.php
@@ -894,6 +894,11 @@ class WP_Theme_JSON_Gutenberg_Test extends WP_UnitTestCase {
 									'background' => 'blue',
 								),
 							),
+							':seen'  => array(
+								'color' => array(
+									'background' => 'ivory',
+								),
+							),
 						),
 					),
 				),